--- a/kll_test.go
+++ b/kll_test.go
@@ -2,7 +2,9 @@ package kll
 
 import (
 	"math/rand"
+	"sort"
 	"testing"
+	"time"
 )
 
 func benchmarkAdd(b *testing.B, cons func() float64, k int) {
@@ -39,4 +41,21 @@ func BenchmarkAddNormal_100(b *testing.B) {
 
 func BenchmarkAddNormal_1000(b *testing.B) {
 	benchmarkAdd(b, rand.NormFloat64, 1000)
+}
+
+func TestCompactorInsertionSort(t *testing.T) {
+	rng := rand.New(rand.NewSource(time.Now().UnixNano()))
+
+	for i := 0; i < 10000; i++ {
+		l := rng.Intn(1000)
+		c := make(compactor, l)
+		for i := range c {
+			c[i] = rng.NormFloat64()
+		}
+
+		c.insertionSort()
+		if !sort.Float64sAreSorted([]float64(c)) {
+			t.Fatal("failed to sort: %v", c)
+		}
+	}
 }