--- a/kll_test.go
+++ b/kll_test.go
@@ -46,16 +46,30 @@ func BenchmarkAddNormal_1000(b *testing.B) {
 func TestCompactorInsertionSort(t *testing.T) {
 	rng := rand.New(rand.NewSource(time.Now().UnixNano()))
 
-	for i := 0; i < 10000; i++ {
-		l := rng.Intn(1000)
-		c := make(compactor, l)
-		for i := range c {
-			c[i] = rng.NormFloat64()
-		}
-
-		c.insertionSort()
-		if !sort.Float64sAreSorted([]float64(c)) {
-			t.Fatal("failed to sort: %v", c)
+	for _, dup := range []bool{false, true} {
+		for _, l := range []int{0, 1, 2, 3, 5, 8, 1 << 5, 1 << 10} {
+			for i := 0; i < 100; i++ {
+				c := make(compactor, l)
+				for i := range c {
+					if dup && i%2 == 1 {
+						c[i] = c[i-1]
+					} else {
+						c[i] = rng.NormFloat64()
+					}
+				}
+				cp := make(compactor, l)
+				copy(cp, c)
+				c.insertionSort()
+				if !sort.Float64sAreSorted([]float64(c)) {
+					t.Fatalf("failed to sort: %v", c)
+				}
+				sort.Float64s(cp)
+				for i, v := range c {
+					if v != cp[i] {
+						t.Fatalf("failed to sort: %f!=%f @%d\nexpected: %v, got %v", cp[i], v, i, cp, c)
+					}
+				}
+			}
 		}
 	}
 }