Here's the overall benchcmp output from the changes. It mostly improves things for small k when more work is done per update due to the buffers being smaller.

```
name              old time/op    new time/op    delta
AddNormal_1-8        291ns ± 2%      55ns ± 1%   -81.09%         (p=0.000 n=12+9)
AddNormal_5-8        288ns ± 1%      55ns ± 1%   -80.89%        (p=0.000 n=12+10)
AddNormal_10-8       284ns ± 2%      56ns ± 2%   -80.40%         (p=0.000 n=12+9)
AddNormal_100-8      240ns ± 1%      75ns ± 2%   -68.96%        (p=0.000 n=12+10)
AddNormal_1000-8     151ns ± 1%     143ns ± 0%    -5.24%         (p=0.000 n=10+8)

name              old alloc/op   new alloc/op   delta
AddNormal_1-8        31.0B ± 0%     0.0B ±NaN%  -100.00%        (p=0.000 n=13+10)
AddNormal_5-8        31.0B ± 0%     0.0B ±NaN%  -100.00%        (p=0.000 n=13+10)
AddNormal_10-8       31.0B ± 0%     0.0B ±NaN%  -100.00%        (p=0.000 n=13+10)
AddNormal_100-8      26.0B ± 0%      5.0B ± 0%   -80.77%        (p=0.000 n=13+10)
AddNormal_1000-8     2.00B ± 0%     2.00B ± 0%      ~     (all samples are equal)
```
<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgryski/go-kll/2)
<!-- Reviewable:end -->