@sasha-s 2016-08-11
maybe update gob serialization for coin?

---

Reviewed 2 of 5 files at r1.
Review status: 2 of 5 files reviewed at latest revision, 1 unresolved discussion.

---

*[coin.go, line 16 \[r1\]](https://reviewable.io:443/reviews/dgryski/go-kll/2#-KOux3NIwwnSvF5zeb65:-KOux3NIwwnSvF5zeb66:by9v4ff) ([raw file](https://github.com/dgryski/go-kll/blob/c67b83da5af289b17e5f3c5d91912673a66acc5e/coin.go#L16)):*
> ```Go
> // to the random number generator.
> type coin struct {
> 	st   uint64
> ```

You could make it smaller with
```
st uint64
mask uint64
```
mask starts with 1
and then shift the mask to the left

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/dgryski/go-kll/2#-:-KOuxSpIuY9nwl5smuFK:b-s2j321)*
<!-- Sent from Reviewable.io -->