@sasha-s 2016-08-11
Reviewed 3 of 5 files at r1.
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/dgryski/go-kll/2)*
<!-- Sent from Reviewable.io -->