@zeebo 2016-08-11
it didn't have that property before (it always used "math/rand".Intn(2) and didn't attempt to save the state).

---

Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[coin.go, line 16 \[r1\]](https://reviewable.io:443/reviews/dgryski/go-kll/2#-KOux3NIwwnSvF5zeb65:-KOvHqLPOmad_yTRzfKi:b-896fix) ([raw file](https://github.com/dgryski/go-kll/blob/c67b83da5af289b17e5f3c5d91912673a66acc5e/coin.go#L16)):*
<details><summary><i>Previously, sasha-s wrote…</i></summary>
> You could make it smaller with
> ```
> st uint64
> mask uint64
> ```
> mask starts with 1
> and then shift the mask to the left
</details>
Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/dgryski/go-kll/2#-:-KOvHsgLOKQ1ZfLFC1K4:bfmce57)*
<!-- Sent from Reviewable.io -->