@sasha-s 2016-08-11
What would happen if someone does
```
var s kll.Sketch
s.UnmarshalBinary(...)   // or gob.xxxx
```
?
It seems that the coin would be in zero state, which would lead to state being zero forever.

Maybe add newCoin to UnmarshalBinary?

---

Reviewed 2 of 2 files at r2.
Review status: all files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/dgryski/go-kll/2#-:-KOvIxO7NEfjtqPwYO2b:b-xdyw3o)*
<!-- Sent from Reviewable.io -->