@zeebo 2016-08-11
Good point. I can either do that or have the coin implementation pick a random state if the state is zero on the call to toss. The latter seems more robust because you don't have to worry about a coin just returning zero forever. What do you think?

---

Review status: :shipit: all files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/dgryski/go-kll/2#-:-KOvKegj1Pfwy6L1rliu:byd3qvq)*
<!-- Sent from Reviewable.io -->