@sasha-s 2016-08-11
I am fine with either.
As Damien pointed out, it should not get into a zero state from a non-zero state.
on the other hand, an extra if in xorshift would most certainly  have a negligible impact on performance and it looks a bit cleaner.

---

Review status: :shipit: all files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/dgryski/go-kll/2#-:-KOvLI7mWHt6CM9a_dmU:b-135z81)*
<!-- Sent from Reviewable.io -->