@sasha-s 2016-08-11
Reviewed 3 of 3 files at r3.
Review status: all files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/dgryski/go-kll/2)*
<!-- Sent from Reviewable.io -->