@sasha-s 2016-08-11
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: :shipit: all files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/dgryski/go-kll/2#-:-KOvORUnhk8_rDUisiLo:bnfp4nl)*
<!-- Sent from Reviewable.io -->