@sasha-s 2016-08-11
Reviewed 2 of 2 files at r4.
Review status: 4 of 6 files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/dgryski/go-kll/2)*
<!-- Sent from Reviewable.io -->